Create Serialization and non serialization binary trees
When we want to save a binary tree as a string in a certain format according to a certain traversal method, the binary tree established in memory can be preserved permanently. Serialization can be modified based on binary tree traversal of first order, middle order, second order and sequence. The result of serialization is a string. When serializing, empty nodes (ා) are represented by some symbol, and then! Represents the end of a node value (value!).